+++ sdk/python/pulumi_azurerm/cache/v20190701/outputs.py
@@ -5,7 +5,7 @@
 import warnings
 import pulumi
 import pulumi.runtime
-from typing import Any, Dict, List, Mapping, Optional, Tuple, Union
+from typing import Any, Mapping, Optional, Sequence, Union
 from ... import _utilities, _tables
 
 __all__ = [
@@ -59,16 +59,16 @@ class RedisInstanceDetailsResponse(dict):
     """
     def __init__(__self__, *,
                  is_master: bool,
-                 non_ssl_port: float,
-                 shard_id: float,
-                 ssl_port: float,
+                 non_ssl_port: int,
+                 shard_id: int,
+                 ssl_port: int,
                  zone: str):
         """
         Details of single instance of redis.
         :param bool is_master: Specifies whether the instance is a master node.
-        :param float non_ssl_port: If enableNonSslPort is true, provides Redis instance Non-SSL port.
-        :param float shard_id: If clustering is enabled, the Shard ID of Redis Instance
-        :param float ssl_port: Redis instance SSL port.
+        :param int non_ssl_port: If enableNonSslPort is true, provides Redis instance Non-SSL port.
+        :param int shard_id: If clustering is enabled, the Shard ID of Redis Instance
+        :param int ssl_port: Redis instance SSL port.
         :param str zone: If the Cache uses availability zones, specifies availability zone where this instance is located.
         """
         pulumi.set(__self__, "is_master", is_master)
@@ -87,7 +87,7 @@ class RedisInstanceDetailsResponse(dict):
 
     @property
     @pulumi.getter(name="nonSslPort")
-    def non_ssl_port(self) -> float:
+    def non_ssl_port(self) -> int:
         """
         If enableNonSslPort is true, provides Redis instance Non-SSL port.
         """
@@ -95,7 +95,7 @@ class RedisInstanceDetailsResponse(dict):
 
     @property
     @pulumi.getter(name="shardId")
-    def shard_id(self) -> float:
+    def shard_id(self) -> int:
         """
         If clustering is enabled, the Shard ID of Redis Instance
         """
@@ -103,7 +103,7 @@ class RedisInstanceDetailsResponse(dict):
 
     @property
     @pulumi.getter(name="sslPort")
-    def ssl_port(self) -> float:
+    def ssl_port(self) -> int:
         """
         Redis instance SSL port.
         """
@@ -153,12 +153,12 @@ class ScheduleEntryResponse(dict):
     """
     def __init__(__self__, *,
                  day_of_week: str,
-                 start_hour_utc: float,
+                 start_hour_utc: int,
                  maintenance_window: Optional[str] = None):
         """
         Patch schedule entry for a Premium Redis Cache.
         :param str day_of_week: Day of the week when a cache can be patched.
-        :param float start_hour_utc: Start hour after which cache patching can start.
+        :param int start_hour_utc: Start hour after which cache patching can start.
         :param str maintenance_window: ISO8601 timespan specifying how much time cache patching can take. 
         """
         pulumi.set(__self__, "day_of_week", day_of_week)
@@ -176,7 +176,7 @@ class ScheduleEntryResponse(dict):
 
     @property
     @pulumi.getter(name="startHourUtc")
-    def start_hour_utc(self) -> float:
+    def start_hour_utc(self) -> int:
         """
         Start hour after which cache patching can start.
         """
@@ -200,12 +200,12 @@ class SkuResponse(dict):
     SKU parameters supplied to the create Redis operation.
     """
     def __init__(__self__, *,
-                 capacity: float,
+                 capacity: int,
                  family: str,
                  name: str):
         """
         SKU parameters supplied to the create Redis operation.
-        :param float capacity: The size of the Redis cache to deploy. Valid values: for C (Basic/Standard) family (0, 1, 2, 3, 4, 5, 6), for P (Premium) family (1, 2, 3, 4, 5).
+        :param int capacity: The size of the Redis cache to deploy. Valid values: for C (Basic/Standard) family (0, 1, 2, 3, 4, 5, 6), for P (Premium) family (1, 2, 3, 4, 5).
         :param str family: The SKU family to use. Valid values: (C, P). (C = Basic/Standard, P = Premium).
         :param str name: The type of Redis cache to deploy. Valid values: (Basic, Standard, Premium)
         """
@@ -215,7 +215,7 @@ class SkuResponse(dict):
 
     @property
     @pulumi.getter
-    def capacity(self) -> float:
+    def capacity(self) -> int:
         """
         The size of the Redis cache to deploy. Valid values: for C (Basic/Standard) family (0, 1, 2, 3, 4, 5, 6), for P (Premium) family (1, 2, 3, 4, 5).
         """
